refactor: add module for AuthGuard
