+++ src/album/album.module.ts
@@ -2,11 +2,11 @@ import { Module } from '@nestjs/common';
 import { AlbumService } from './album.service';
 import { AlbumController } from './album.controller';
 import { DatabaseService } from 'src/database/database.service';
-import { TokenModule } from 'src/token/token.module';
+import { GuardModule } from 'src/guard/guard.module';
 
 @Module({
   providers: [AlbumService, DatabaseService],
   controllers: [AlbumController],
-  imports: [TokenModule],
+  imports: [GuardModule],
 })
 export class AlbumModule {}
