+++ src/artist/artist.module.ts
@@ -2,11 +2,11 @@ import { Module } from '@nestjs/common';
 import { ArtistController } from './artist.controller';
 import { ArtistService } from './artist.service';
 import { DatabaseService } from 'src/database/database.service';
-import { TokenModule } from 'src/token/token.module';
+import { GuardModule } from 'src/guard/guard.module';
 
 @Module({
   controllers: [ArtistController],
   providers: [ArtistService, DatabaseService],
-  imports: [TokenModule],
+  imports: [GuardModule],
 })
 export class ArtistModule {}
