+++ src/favs/favs.module.ts
@@ -2,11 +2,11 @@ import { Module } from '@nestjs/common';
 import { FavsService } from './favs.service';
 import { FavsController } from './favs.controller';
 import { DatabaseService } from 'src/database/database.service';
-import { TokenModule } from 'src/token/token.module';
+import { GuardModule } from 'src/guard/guard.module';
 
 @Module({
   providers: [FavsService, DatabaseService],
   controllers: [FavsController],
-  imports: [TokenModule],
+  imports: [GuardModule],
 })
 export class FavsModule {}
