+++ src/token/token.module.ts
@@ -1,10 +1,11 @@
 import { Module } from '@nestjs/common';
 import { TokenService } from './token.service';
-import { ConfigService } from '@nestjs/config';
-import { JwtService } from '@nestjs/jwt';
+import { ConfigModule, ConfigService } from '@nestjs/config';
+import { JwtModule } from '@nestjs/jwt';
 
 @Module({
-  providers: [JwtService, ConfigService, TokenService],
-  exports: [TokenService],
+  providers: [TokenService, ConfigService],
+  exports: [TokenService, ConfigService],
+  imports: [ConfigModule, JwtModule],
 })
 export class TokenModule {}
