+++ src/track/track.module.ts
@@ -2,11 +2,11 @@ import { Module } from '@nestjs/common';
 import { TrackService } from './track.service';
 import { TrackController } from './track.controller';
 import { DatabaseService } from 'src/database/database.service';
-import { TokenModule } from 'src/token/token.module';
+import { GuardModule } from 'src/guard/guard.module';
 
 @Module({
   providers: [TrackService, DatabaseService],
   controllers: [TrackController],
-  imports: [TokenModule],
+  imports: [GuardModule],
 })
 export class TrackModule {}
