+++ src/user/user.module.ts
@@ -3,11 +3,11 @@ import { UserController } from './user.controller';
 import { UserService } from './user.service';
 import { DatabaseService } from 'src/database/database.service';
 import { ConfigService } from '@nestjs/config';
-import { TokenModule } from 'src/token/token.module';
+import { GuardModule } from 'src/guard/guard.module';
 
 @Module({
   controllers: [UserController],
   providers: [UserService, DatabaseService, ConfigService],
-  imports: [TokenModule],
+  imports: [GuardModule],
 })
 export class UserModule {}
